+++ changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
 ## Upcoming
+* Zombies can now mine. 5% chance to spawn with the miner AI with which they can break blocks to reach the player. If they have the right tools in the *Off Hand* then they'll mine faster.
 * Fixed Mobs running away from explosion seemingly losing AI if the target died
 * Creepers will no longer try to breach when in water
+* Creepers now stop the explosion when breaching if the target moves away.
+* Creepers now stop the explosion when launching if the target dies
+* Reduced avoid speed of explosions
 
 ## Alpha 1.0.2
 * Creepers now Swell from the correct distance